organize components a bit better
