+++ src/components/Battle/EnemyGroupSection.tsx
@@ -2,8 +2,8 @@ import { useContext } from 'react';
 import styled from 'styled-components';
 
 import EnemyGroup from './EnemyGroup';
-import { AppStateContext } from '../state';
-import { LEFT_ENEMY_GROUP, RIGHT_ENEMY_GROUP } from '../constants';
+import { AppStateContext } from '../../state';
+import { LEFT_ENEMY_GROUP, RIGHT_ENEMY_GROUP } from '../../constants';
 
 const EnemySection = styled.section`
   flex: 0 1 30%;
