+++ src/components/Battle/Hero.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 
-import { EntityType } from '../types';
-import Window from './Window';
+import { EntityType } from '../../types';
+import Window from '../Window';
 
 const HeroContainer = styled(Window)`
   display: flex;
