+++ src/components/Battle/index.ts
@@ -1,0 +1,3 @@
+import Battle from './Battle';
+
+export default Battle;
