--- src/components/Enemies.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Enemies = () => (
-  <h1>Enemy names, drops, attributes and etc to be worked out here</h1>
-);
-
-export default Enemies;
