--- src/components/Player.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Player = () => (
-  <h1>Player name, attributes, inventory and etc go here</h1>
-);
-
-export default Player;
